--All Style
-Refactored styled components   -README file added path to the dir in the installation
